+++ app/javascript/components/DataStory/constants.js
@@ -1,6 +1,5 @@
 export const palette = {
   indirect: "#2e3348",
   mediating: "#437387",
-  // outcome: "#84b65b",
   outcome: "#1b6002",
 };
