+++ app/javascript/packs/data_story.js
@@ -1,6 +1,3 @@
-// This logic renders a live-updating preview of a layer's widget on the layer
-// form partial.
-
 import React from "react";
 import ReactDOM from "react-dom";
 import { DataStoryRoot } from "components/DataStory/Root";
